backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
grpcio
beautifulsoup4
